move helper to common component
