+++ src/views/schedule-details/schedule-details-runs-chart-glyph/schedule-details-runs-chart-glyph.tsx
@@ -7,8 +7,9 @@ import { MdHistory } from 'react-icons/md';
 
 import { WORKFLOW_STATUSES } from '@/views/shared/workflow-status-tag/workflow-status-tag.constants';
 
-import getChartGlyphColor from './helpers/get-chart-glyph-color';
-import getChartStatusIcon from './helpers/get-chart-status-icon';
+import getChartGlyphColor from '../schedule-details-runs-chart/helpers/get-chart-glyph-color';
+import getChartStatusIcon from '../schedule-details-runs-chart/helpers/get-chart-status-icon';
+
 import {
   CHART_GLYPH_MARKER_SIZE_PX,
   CHART_GLYPH_TEST_IDS,
