+++ src/views/schedule-details/schedule-details-runs-chart/helpers/__tests__/get-chart-glyph-color.test.ts
