+++ src/views/schedule-details/schedule-details-runs-chart/helpers/__tests__/get-chart-status-icon.test.ts
