+++ src/views/schedule-details/schedule-details-runs-chart/helpers/get-chart-glyph-color.ts
@@ -2,7 +2,7 @@ import { type Theme } from 'baseui';
 
 import { WORKFLOW_STATUSES } from '@/views/shared/workflow-status-tag/workflow-status-tag.constants';
 
-import { type ChartGlyphVariant } from '../schedule-details-runs-chart-glyph.types';
+import { type ChartGlyphVariant } from '../../schedule-details-runs-chart-glyph/schedule-details-runs-chart-glyph.types';
 
 export default function getChartGlyphColor(
   theme: Theme,
